Get-DattoApplication -daysUntil input 0 value
The endpoint accept a '0' value, which is different from not including the parameter, by default with no parameter set the API returns 10 days of historical data.

With 0 value, only remote IDs can be returned if chosen by the module user. Since the API accepts the '0' value we should include it as well as valid input.